Model routers instead of having this sketchy, throwaway node
It was getting weird to pass this node with a  null address around just so I could bootstrap off of it.

Instead, it should be treated like a "router." A router is a node whose address we dgaf about, because we just send messages to it to bootstrap off of.

They will get chained onto the end of the nodes from neartest_to in the case that we don't have enough nodes already. This frees me up to quit thinking about how to make special queries to this node. Instead, the routing table just knows about it, and I get to ask the routing table for nodes just like usual.