Setup Network module to be able to send and receive packets
The idea is to wrap this network module up so that it is a general-purpose async networking library. I'm not certain of the guarantees it offers right now, except that it will

- Maintain a health connection to the rest of the network
- Try to deliver packets as close as possible to their destination
- Emit events when it receives a packet

It's up to the client to determine whether it wants to relay packets or whatever given the application's needs.