Move stdin message sending out of client model
First of all, Client is a bad name for what this module is. It's really the chat functionality that a client needs to interface with.

Since I'm trying to build a C API and ultimately a Cocoa client, I need to extract all this implementation detail out (main.rs is like the poor mans client implementation using stdio) and leave a pure command/event interface.