Implement a shutdown procedure
Adds Shutdown command and event to both the Network and the Client

Sending a Shutdown command to the client causes it to send on to the Network, which hangs up all of it's listening ports.

Upon the ports closing, the Network stops its event loop and emits a Shutdown event, which in turn causes the Client to do the same.